requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
boto3==1.19.12
botocore==1.22.12
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
Flask==1.1.1
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
hawk-auth==0.1.0rc1634278026
hawk-core==0.1.0rc1634278027
hawk-db==0.1.0rc1634278028
hawk-models==0.1.0rc1634278029
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
jmespath==0.10.0
JSON-log-formatter==0.3.0
MarkupSafe==2.0.1
marshmallow==3.9.1
mypy-extensions==0.4.3
numpy==1.21.4
oauthlib==3.1.1
pandas==1.3.4
pika==1.1.0
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==3.0.6
pyserde==0.3.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.3.20
stringcase==1.2.0
summ-common==0.1.0rc1633590179
summ-logging==0.1.0rc1633590180
summ-rabbitmq==0.1.0rc1633590178
summ-web==0.1.0rc1633590181
typing-extensions==3.7.4.3
typing-inspect==0.7.1
uritemplate==4.1.1
urllib3==1.26.7
webargs==6.1.1
Werkzeug==2.0.2
